+++ internal/rooms/participant.go
@@ -18,7 +18,7 @@ type Participant struct {
 	IsMicroOn    bool        `json:"isMicroOn"`
 	IsSpeakerOn  bool        `json:"isSpeakerOn"`
 	CameraType   *string     `json:"cameraType"`
-	BatteryLife  int64       `json:"batteryLife"`
+	BatteryLife  float64     `json:"batteryLife"`
 }
 
 func (p *Participant) String() string {
